+++ NEWS.md
@@ -95,6 +95,8 @@ upon client creation.
 
 * Ensured a clearer error message when token response is in unexpected format.
 
+* Ensured a clearer error when retrieved state store entry is in unexpected format.
+
 * Ensured a clearer error message when retrieved userinfo cannot be parsed as JSON.
 
 * Immediate error when `OAuthProvider` uses `HS*` algorithm but